requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
azure-ai-vision-imageanalysis==1.0.0
    # via -r requirements.in
azure-core==1.32.0
    # via azure-ai-vision-imageanalysis
bidict==0.23.1
    # via python-socketio
blinker==1.9.0
    # via flask
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via flask
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via eventlet
eventlet==0.38.2
    # via -r requirements.in
ffmpeg-python==0.2.0
    # via -r requirements.in
flask==3.1.0
    # via -r requirements.in
future==1.0.0
    # via ffmpeg-python
greenlet==3.1.1
    # via eventlet
h11==0.14.0
    # via
    #   httpcore
    #   wsproto
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
isodate==0.7.2
    # via azure-ai-vision-imageanalysis
itsdangerous==2.2.0
    # via flask
jinja2==3.1.5
    # via flask
jiter==0.8.2
    # via openai
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
openai==1.59.5
    # via -r requirements.in
pydantic==2.10.5
    # via openai
pydantic-core==2.27.2
    # via pydantic
pygments==2.19.1
    # via rich
python-dotenv==1.0.1
    # via -r requirements.in
python-engineio==4.11.2
    # via python-socketio
python-socketio==5.12.1
    # via -r requirements.in
requests==2.32.3
    # via azure-core
rich==13.9.4
    # via -r requirements.in
simple-websocket==1.1.0
    # via python-engineio
six==1.17.0
    # via azure-core
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   azure-ai-vision-imageanalysis
    #   azure-core
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.3.0
    # via requests
websockets==14.1
    # via -r requirements.in
werkzeug==3.1.3
    # via flask
wsproto==1.2.0
    # via simple-websocket
